requirements.txt
selenium
webdriver-manager
pandas
openpyxl
schedule
python-dotenv
